requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.7
APScheduler>=3.10.4
python-dotenv>=1.0.0
lxml>=4.9.3
